+++ pages/notifications.php
@@ -106,11 +106,11 @@ $table->set_sql('*', "{block_advnotifications}", "deleted = 0");
 
 // Add navigation controls before the table.
 echo '<div id="advnotifications_manage">
-        <a class="btn instance" href="' .
+        <a class="btn btn-secondary instance" href="' .
             $CFG->wwwroot . '/blocks/advnotifications/pages/restore.php' . $param . '">' .
             get_string('advnotifications_nav_restore', 'block_advnotifications') .
         '</a>&nbsp;&nbsp;
-        <a class="btn instance" href="' .
+        <a class="btn btn-secondary instance" href="' .
             $CFG->wwwroot . '/admin/settings.php?section=blocksettingadvnotifications' . $xparam . '">' .
             get_string('advnotifications_nav_settings', 'block_advnotifications') .
         '</a><br><br>